fix: resolve detekt LongMethod and TooManyFunctions violations
- Extract buildByModeStats from buildGoalStats (GoalWorkflowStatsSupport)
- Inline single-line ensureGoalRunSessionsColumns (DatabaseColumnMigrations)
- Split goal prose handlers into McpGoalToolHandlers to stay under file function limit
- Extract seedGoalStatsDb helper in CliRuntimeTest to shorten test function
